// apps/frontend/core-dapp/src/types/market.types.ts
export type Range = [number, number]

export type SupportedMarketID = 'prefakestock' | 'prefaketoken'

export type MarketType = 'preICO' | 'preIPO'

export type Market = {
  iconName: IconName
  type: MarketType
  name: string
  urlId: SupportedMarketID
  companyName: string
  address: SupportedMarkets
  long: {
    tokenAddress: SupportedMarketTokens
    poolAddress: SupportedMarketPools
  }
  short: {
    tokenAddress: SupportedMarketTokens
    poolAddress: SupportedMarketPools
  }
  static: {
    valuationRange: Range
  }
}